request proof with interval

+++ ariesframework/src/main/java/org/hyperledger/ariesframework/anoncreds/formats/AnoncredsCredentialFormatService.kt
@@ -381,7 +381,7 @@ class AnoncredsCredentialFormatService(
             revocationStatusList = AnonCredsObjects.fetchRevocationStatusList(
                 agent,
                 revocationRegistryDefinitionId,
-                dateToTimestamp(Date()),
+                dateToTimestamp(Date()).toULong(),
             )
         }
